refactor: project 관련 컴포넌트 shared로 이동
단순  UI 컴포넌트
