+++ src/features/index.ts
@@ -1,3 +1,1 @@
-export * from '../shared/ui/profile';
-export * from './project';
 export * from './signup';
