--- src/features/project/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as ProjectCard } from './projectCard/ProjectCard.tsx';
