+++ src/shared/ui/project/index.ts
