+++ src/shared/ui/project/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as ProjectCard } from './projectCard/ProjectCard.tsx';
+export { default as ProjectCardList } from './projectCardList/ProjectCardList.tsx';
