+++ src/shared/ui/project/ui/projectCard/ProjectCard.styles.ts
