+++ src/shared/ui/project/ui/projectCardList/ProjectCardList.styles.ts
